- made menu items pulse when the mouse/joystick goes over them
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk@174 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ tuxkart/src/BaseGUI.h
@@ -1,4 +1,4 @@
-//  $Id: BaseGUI.h,v 1.1 2004/08/04 16:34:31 jamesgregory Exp $
+//  $Id: BaseGUI.h,v 1.2 2004/08/05 16:47:18 jamesgregory Exp $
 //
 //  TuxKart - a fun racing game with go-kart
 //  Copyright (C) 2004 Steve Baker <sjbaker1@airmail.net>
@@ -35,7 +35,7 @@ public:
 	BaseGUI() {}
 	virtual ~BaseGUI() {}
 	
-	virtual void update(int dt) = 0;
+	virtual void update(float dt) = 0;
 	virtual void click(int button, int x, int y) = 0;
 	virtual void point(int x, int y) = 0; 
 	virtual void stick(int x, int y) = 0;
